Move Slack message creation to blocks.ts
